PresentationFramework: Restore baml loading code for themes.
Since we have no way to build .baml files, we originally hacked
the theme loading code to look for .xaml files, so we could build
themes without a .baml build process. It turns out that
applications can ship with their own themes, and we still need to
be able to load baml in these cases.

For https://github.com/ValveSoftware/Proton/issues/4983

Unfortunately there are more bugs affecting the launcher.